Check if adapter is already loaded
I adapter is already loaded, it's not necesary to try to charge it again. Or if it's not available at root, trying to load it as it will cause a fatal error. It is probably most pertinent if you can let the choice to the user of the place of the script