External token authentication and authorization
The traditional Janus token authentication provides limited ability to control
access. It allows blocking access to plugins entirely, but no finer-grained
control than that, e.g., it is not possible to have a token that allows access
to particular rooms from the videoroom and/or particular streams from the
streaming plugin.

With this patch, external authentication is allowed. In the default case,
authentication is accomplished via an HTTP server. An example server
implementation is provided in the new directory auth-external, which
demonstrates how to allow certain requests while denying others, even to the
same plugin.

Although the default external authenticator sends its requests to an HTTP
server, other authenticators could be written. Only a single function pointer
need be changed to switch to a different authenticator.

The HTTP authenticator uses libcurl to talk with the authenticator server. If
libcurl is not available and external authentication is requested, a default
authenticator that always disallows requests is used.

The general->token_auth configuration key in janus.cfg may be set to a boolean
value as before ("yes", "no", "true", "false"). It may now additionally be set
to "external" to enable the external token authenticator.

One new configuration key is added to janus.cfg: general->token_auth_private.
For the default external authenticator (sending HTTP requests), this private
data is set to the URL at which the authenticator is listening.