Fix SSL library detection
Currently, searching for RAND_egd doesn't seem to work properly against
OpenSSL 1.1, as it always returns false.

Instead, searching for the function tls_config_set_ca_mem in libtls seems to
be the preferred way to determine if LibreSSL is being compiled against, as
this function does not exist in OpenSSL.

Openntpd uses this callout in their configure.ac. Because Openntpd is
maintained by the OpenBSD foundation, and they also maintain LibreSSL, it would
be a good idea to follow their example.