Reduce keepalive time to less then 30 seconds
We have encountered environments where the WebSocket connection is closed after 30 seconds of inactivity. The current Janus keepalive is set to 30 seconds so we have a race condition and we often see the Jauns WebSocket closed in these environments. I have changed this to 25 seconds as this is the default that Socket IO uses (pingInterval):

https://socket.io/docs/server-api/

This change solves the issue for us so I'm proposing it for everyone in case others come across similar environments.