add padding for sdes cname
sdes chunk should end with null terminator

see: https://tools.ietf.org/html/rfc1889
6.4 SDES: Source description RTCP packet
 Items are contiguous, i.e., items are not individually padded to a
32-bit boundary. Text is not null terminated because some multi-octet
encodings include null octets. The list of items in each chunk is
terminated by one or more null octets, the first of which is
interpreted as an item type of zero to denote the end of the list,
and the remainder as needed to pad until the next 32-bit boundary. A
chunk with zero items (four null octets) is valid but useless.