nosip: better port choice (fixed number of attempts replaced with a full scan within a given range)
Test case:
min_port = 10000
max_port = 60000
Ports from 10000 to 59995 are opened on a first session start and remained open
Port slider set to 10000 (to start discovering from 10000 as a worst case)

It takes 30-40ms to run to port 5996 for Ubuntu laptop with core-i5@2.5GHz.
It takes ~150ms if we close rtp fd on unsuccessful bind as it was originally done.
One can implement look-up table (as an array of bit-masks) to pick up free port for 0ms.