Session media attributes has_audio and has_video are not cleared when call terminates and are carried over into the next call. This sometimes causes the subsequent calls to behave unpredictably. Added code resettings these attributes to janus_sip_media_cleanup().
Also, analysed other media attributes and found some of them are not cleaned up when
a call terminates. Additionally, some attributes are not initialiazed when session is
allocated. Added code to reset media attributes that otherwise are not guaranteed
to have initial values when a call terminates or when a session is allocated.