Fixed handling of offerless reINVITE.
"Update" request is generated by client in two cases:
1. when initiating a session update, sending a re-INVITE request
2. when handling a session update, sending a re-INVITE response
There is a problem with the second case. When a re-INVITE with SDP offer
is received everything is fine. However, if an offerless re-INVITE is
received it is responded to with a re-IVITE request, which is wrong. The
problem seems to be that the "update" request with sdp of type "offer"
is hard-wired with sending a re-INVITE request and sdp of "answer" is
hard-wired with sending a re-INVITE response. That obviously doesn't
work with offerless re-INVITE.

To fix the problem the re-INVITE response is sent when there is a
re-INVITE request is in progress, otherwise, a new re-INVITE request is
sent. The decision is based on the state of the session that now
recognises client or server re-INVITE progress. Accordingly, two
new states are introduced and maintained:
janus_sip_call_status_incall_reinviting and
janus_sip_call_status_incall_reinvited as extension to the existing
state janus_sip_call_status_incall. Additionally, having the new states
is helpful in preventing "glare" type of scenarios, for example, where
client "update" request collides with SIP re-INVITE request.

For reliability, the new states transitions are driven by nua state
events "ready", "calling" or when receiving a re-INVITE request.

Also, added method update_session_status() that logs session state
changes to assist debugging.

Handling of some client request types such as "hangup", "update",
"hold", "unhold", "info", "dtmf_info" are allowed in session state
janus_sip_call_status_inviting that seems to be a copy-and-paste error.
Have corrected this as well.