Fix media not enabled by subscribers on publisher renegotiations
Since 493632d1aaad when a subscriber is configured the "audio", "video"
and "data" flags are honoured. If they are disabled they will stay
disabled, and if they are enabled then the result will depend on whether
the publisher is sending them and the subscriber has offered to receive
them. The value of the flags will be either the value given in the
configure request or the value they had already if it is not explicitly
given.

When the SDP is updated for a publisher the publisher sends a configure
request to update all its subscribers. Before 493632d1aaad if a
publisher started to publish audio, video or data the audio, video or
data was automatically enabled by the subscribers. However, after that
change the audio, video or data is no longer automatically enabled; if
the audio, video or data was not being sent by the publisher when the
subscriber was created the corresponding flag is disabled, so when the
publisher starts to send it it will stay disabled in the subscriber, as
it was not possible to differentiate between explicitly disabled and
implicitly disabled.

To solve that additional variables are introduced to keep track of
whether the audio, video or data were explicitly configured or not and,
if they were, with which value. This is now taken into account when a
subscriber is updated, so the audio, video or data will be automatically
enabled if the publisher starts to send them and they were not
explicitly disabled by the subscriber.

Besides that, for consistency the configured value is now also taken
into account when switching publishers for a subscriber; if the audio,
video or data was explicitly disabled it will stay disabled after the
switch.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>